Lambda_Assign_4

+++ src/Assign_4/Assign_4.java
@@ -1,5 +1,5 @@
 package Assign_4;
 
 public interface Assign_4 {
-    long result(int x,int y);
+    long result(long x,long y);
 }
